+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/FactoryExtensions/StoredProcedureFactoryExtension.cs
@@ -1,0 +1,130 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Intent.Engine;
+using Intent.EntityFrameworkCore.Repositories.Api;
+using Intent.Modelers.Domain.Api;
+using Intent.Modelers.Domain.Repositories.Api;
+using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Builder;
+using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.Plugins;
+using Intent.Modules.Common.Templates;
+using Intent.Modules.Constants;
+using Intent.Modules.Entities.Repositories.Api.Templates.EntityRepositoryInterface;
+using Intent.Modules.Entities.Repositories.Api.Templates.RepositoryInterface;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates.CustomRepositoryInterface;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates.Repository;
+using Intent.Plugins.FactoryExtensions;
+using Intent.RoslynWeaver.Attributes;
+using Intent.Templates;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.FactoryExtension", Version = "1.0")]
+
+namespace Intent.Modules.EntityFrameworkCore.Repositories.FactoryExtensions
+{
+    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    public class StoredProcedureFactoryExtension : FactoryExtensionBase
+    {
+        public override string Id => "Intent.EntityFrameworkCore.Repositories.StoredProcedureFactoryExtension";
+
+        [IntentManaged(Mode.Ignore)]
+        public override int Order => 0;
+
+        protected override void OnAfterTemplateRegistrations(IApplication application)
+        {
+            var repositoryModels = application.MetadataManager.Domain(application).GetRepositoryModels();
+
+            var dataContractResults = repositoryModels
+                .SelectMany(StoredProcedureHelpers.GetStoredProcedureModels)
+                .Select(x => x.TypeReference?.Element.AsDataContractModel())
+                .Where(x => x != null)
+                .Distinct()
+                .ToArray();
+
+            if (dataContractResults.Any() &&
+                TryGetTemplate<ICSharpFileBuilderTemplate>(application, TemplateFulfillingRoles.Infrastructure.Data.DbContext, out var dbContextTemplate))
+            {
+                dbContextTemplate.CSharpFile.OnBuild(file =>
+                {
+                    var @class = file.Classes.Single();
+                    var onModelCreating = @class.Methods.Single(x => x.Name == "OnModelCreating");
+
+                    foreach (var dc in dataContractResults)
+                    {
+                        var typeName = dbContextTemplate.GetTypeName(TemplateFulfillingRoles.Domain.DataContract, dc);
+
+                        @class.AddProperty(
+                            type: $"DbSet<{typeName}>",
+                            name: typeName.Pluralize(),
+                            configure: prop => prop.AddMetadata("model", dc));
+
+                        onModelCreating.AddStatement($"modelBuilder.Entity<{typeName}>().HasNoKey().ToView(null);");
+                    }
+                });
+            }
+
+            var classRepositories = repositoryModels
+                .Where(x => x.TypeReference.Element.IsClassModel())
+                .Select(x => (Entity: x.TypeReference.Element.AsClassModel(), Repository: x));
+
+            foreach (var (entity, repository) in classRepositories)
+            {
+                var storedProcedures = StoredProcedureHelpers.GetStoredProcedureModels(repository);
+                if (!storedProcedures.Any())
+                {
+                    continue;
+                }
+
+                if (TryGetTemplate<ICSharpFileBuilderTemplate>(application, EntityRepositoryInterfaceTemplate.TemplateId, entity, out var interfaceTemplate))
+                {
+                    StoredProcedureHelpers.ApplyInterfaceMethods(interfaceTemplate, storedProcedures);
+                }
+
+                if (TryGetTemplate<RepositoryTemplate>(application, RepositoryTemplate.TemplateId, entity, out var implementationTemplate))
+                {
+                    if (storedProcedures.Any(x => x.TypeReference.Element?.Id != implementationTemplate.Model.Id))
+                    {
+                        implementationTemplate.CSharpFile.AfterBuild(file =>
+                        {
+                            var @class = file.Classes.First();
+
+                            var parameters = @class.Constructors
+                                .SelectMany(x => x.Parameters)
+                                .Where(x => x.Name == "dbContext");
+
+                            foreach (var parameter in parameters)
+                            {
+                                parameter.IntroduceReadonlyField();
+                            }
+                        });
+                    }
+
+                    StoredProcedureHelpers.ApplyImplementationMethods(implementationTemplate, storedProcedures);
+                }
+            }
+        }
+
+        private bool TryGetTemplate<T>(
+            IApplication application,
+            string templateId,
+            object model,
+            out T template) where T : class
+        {
+            template = application.FindTemplateInstance<T>(templateId, model);
+            return template != null;
+        }
+
+        private bool TryGetTemplate<T>(
+            IApplication application,
+            string templateId,
+            out T template) where T : class
+        {
+            template = application.FindTemplateInstance<T>(templateId);
+            return template != null;
+        }
+    }
+}