+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/Templates/TemplateExtensions.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates.CustomRepository;
+using Intent.Modules.EntityFrameworkCore.Repositories.Templates.CustomRepositoryInterface;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.PagedList;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.Repository;
 using Intent.Modules.EntityFrameworkCore.Repositories.Templates.RepositoryBase;
@@ -13,6 +15,25 @@ namespace Intent.Modules.EntityFrameworkCore.Repositories.Templates
 {
     public static class TemplateExtensions
     {
+        public static string GetCustomRepositoryName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Domain.Repositories.Api.RepositoryModel
+        {
+            return template.GetTypeName(CustomRepositoryTemplate.TemplateId, template.Model);
+        }
+
+        public static string GetCustomRepositoryName(this IIntentTemplate template, Intent.Modelers.Domain.Repositories.Api.RepositoryModel model)
+        {
+            return template.GetTypeName(CustomRepositoryTemplate.TemplateId, model);
+        }
+
+        public static string GetCustomRepositoryInterfaceName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Domain.Repositories.Api.RepositoryModel
+        {
+            return template.GetTypeName(CustomRepositoryInterfaceTemplate.TemplateId, template.Model);
+        }
+
+        public static string GetCustomRepositoryInterfaceName(this IIntentTemplate template, Intent.Modelers.Domain.Repositories.Api.RepositoryModel model)
+        {
+            return template.GetTypeName(CustomRepositoryInterfaceTemplate.TemplateId, model);
+        }
         public static string GetPagedListName(this IIntentTemplate template)
         {
             return template.GetTypeName(PagedListTemplate.TemplateId);